Transfered widget to npm generated and modularized ntc library
