+++ README.md
@@ -1,36 +1,43 @@
-Below are the steps to get your widget running. You can also find instructions at:
+# @figma/create-widget
 
-https://www.figma.com/widget-docs/setup-guide/
+This repo was created by @figma/create-widget
 
-This widget template uses Typescript and NPM, two standard tools in creating JavaScript applications.
+## Getting started with widget development
 
-First, download Node.js which comes with NPM. This will allow you to install TypeScript and other
-libraries. You can find the download link here:
+Run the following command to start building your widget
 
-https://nodejs.org/en/download/
+```bash
+npm run dev
+```
 
-Next, install TypeScript and the latest type definitions by running:
+1. Log in to your account and open the Figma desktop app
+2. You can open any existing FigJam document or create a new one.
+3. Go to Menu > Widgets > Development > "Import widget from manifest..."
+4. Select the manifest.json in this folder
 
-npm install
+## Organization
 
-If you are familiar with JavaScript, TypeScript will look very familiar. In fact, valid JavaScript code
-is already valid Typescript code.
+This widget uses:
 
-TypeScript adds type annotations to variables. This allows code editors such as Visual Studio Code
-to provide information about the Figma API while you are writing code, as well as help catch bugs
-you previously didn't notice.
+- [esbuild](https://esbuild.github.io/) for bundling
+- [typescript](https://www.typescriptlang.org/) for typechecking
 
-For more information, visit https://www.typescriptlang.org/
+| file/folder   | description                                                                      |
+| ------------- | -------------------------------------------------------------------------------- |
+| manifest.json | The widget's [manifest.json](https://www.figma.com/widget-docs/widget-manifest/) |
+| widget-src/   | Contains the widget code                                                         |
 
-Using TypeScript requires a compiler to convert TypeScript (code.ts) into JavaScript (code.js)
-for the browser to run.
+### `npm run dev`
 
-We recommend writing TypeScript code using Visual Studio code:
+This is the only command you need to run in development. It will start the following processes for you:
 
-1. Download Visual Studio Code if you haven't already: https://code.visualstudio.com/.
-2. Open this directory in Visual Studio Code.
-3. Compile TypeScript to JavaScript: Run the "Terminal > Run Build Task..." menu item,
-   then select "tsc: watch - tsconfig.json". You will have to do this again every time
-   you reopen Visual Studio Code.
+- bundling
+- typechecking
 
-That's it! Visual Studio Code will regenerate the JavaScript file every time you save.
+### `npm run build`
+
+This runs bundling with minification turned on. You should run this command before releasing your widget.
+
+### `npm run test`
+
+This runs typechecking and makes sure that your widget builds without errors.
