+++ package.json
@@ -1,16 +1,24 @@
 {
   "name": "Color-Swatch",
   "version": "1.0.0",
-  "description": "Your Figma widget",
-  "main": "code.js",
+  "description": "Color Swatch for FigJam and Figma",
   "scripts": {
-    "build": "tsc -p tsconfig.json"
+    "test": "npm run tsc && npm run build",
+    "format": "prettier --write .",
+    "tsc": "tsc --noEmit -p widget-src",
+    "build": "npm run bundle -- --minify",
+    "bundle": "esbuild widget-src/code.tsx --bundle --outfile=dist/code.js",
+    "dev": "concurrently -n tsc,build 'npm run tsc -- --preserveWatchOutput --watch' 'npm run bundle -- --watch'"
   },
-  "author": "",
-  "license": "",
+  "author": "Peter Damrongpiriyapong",
+  "license": "MIT License",
+  "dependencies": {},
   "devDependencies": {
     "@figma/plugin-typings": "*",
     "@figma/widget-typings": "*",
-    "typescript": "*"
+    "concurrently": "^6.3.0",
+    "esbuild": "^0.13.5",
+    "prettier": "^2.3.2",
+    "typescript": "^4.4.2"
   }
 }
