+++ widget-src/tsconfig.json
@@ -4,12 +4,8 @@
     "jsxFactory": "figma.widget.h",
     "jsxFragmentFactory": "figma.widget.Fragment",
     "target": "es6",
-    "lib": [
-      "es6"
-    ],
-    "typeRoots": [
-      "./node_modules/@types",
-      "./node_modules/@figma"
-    ]
+    "lib": ["es6"],
+    "strict": true,
+    "typeRoots": ["../node_modules/@figma"]
   }
 }
